requirements.txt
Django==1.11.6
djangorestframework==3.7.0
numpy==1.13.3
pandas==0.20.3
python-dateutil==2.6.1
pytz==2017.2
scikit-learn==0.19.0
scipy==0.19.1
six==1.11.0
sklearn==0.0
